Adds schema exporting and migration tests

+++ app/src/main/java/org/wspcgir/strong_giraffe/model/Weight.kt
@@ -1,6 +1,6 @@
 package org.wspcgir.strong_giraffe.model
 
-data class Weight(val value: Int) {
+data class Weight(val value: Float) {
     override fun toString(): String {
         return value.toString()
     }
